[test] Include extra calls for expensive sampler
